+++ Erebus/source.cpp
@@ -110,11 +110,11 @@ DWORD WINAPI update( LPVOID args )
 		{
 			double deltaTime = counter.getDeltaTime();
 
-			luaBinds.update( data->controls, deltaTime );
+			luaBinds.update( data->controls, (float)deltaTime );
 			data->workQueue->execute();
 
 			for( int i=0; i<data->particleSystems->size(); i++ )
-				data->particleSystems->at(i)->update( deltaTime );
+				data->particleSystems->at(i)->update( (float)deltaTime );
 
 			collisionHandler.checkCollisions();
 
@@ -125,7 +125,7 @@ DWORD WINAPI update( LPVOID args )
 
 			for( int i=0; i<boundAnimations; i++ )
 			{
-				animationData[i].dt = deltaTime;
+				animationData[i].dt = (float)deltaTime;
 				//data->allAnimations[i].update(deltaTime);
 				data->workQueue->add( updateAnimation, &animationData[i] );
 			}
@@ -292,7 +292,7 @@ int main()
 			engine.draw(&camera);
 
 #ifdef _DEBUG
-			assets.checkHotload(deltaTime);
+			assets.checkHotload((float)deltaTime);
 #endif // DEBUG
 		}
 	}
